pyproject.toml
[tool.poetry]
name = "fastapi-todo"
version = "0.1.0"
description = ""
authors = ["Aneeq Tahir <aneeqtaslam391@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "0.28.0"
psycopg2-binary = "2.9.9"
passlib = "1.7.4"
python-multipart = "0.0.9"
sqlmodel = "0.0.16"
pytest = "8.1.1"
httpx = "0.27.0"
fastapi = "0.110.0"
python-jose = {extras = ["cryptography"], version = "3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
